EWM-233. Splash screen flow
