+++ lib/core/app_build_type.dart
@@ -1,0 +1,5 @@
+enum AppBuildType {
+  development,
+  staging,
+  production,
+}
