+++ lib/feature/no_internet/no_internet_screen_wm.dart
@@ -1,9 +1,11 @@
+import 'package:app/app/router/app_route.dart';
 import 'package:app/core/error_handler_factory.dart';
 import 'package:app/core/wm/custom_wm.dart';
 import 'package:app/feature/no_internet/no_internet_screen.dart';
 import 'package:app/feature/no_internet/no_internet_screen_model.dart';
 import 'package:elementary/elementary.dart';
 import 'package:flutter/widgets.dart';
+import 'package:go_router/go_router.dart';
 import 'package:ui_components_lib/ui_components_lib.dart';
 import 'package:ui_components_lib/v2/text_styles_v2.dart';
 
@@ -33,5 +35,7 @@ class NoInternetScreenWidgetModel
 
   ColorsPaletteV2 get colors => _theme.colors;
 
-  void onPressedTryAgain() {}
+  void onPressedTryAgain() {
+    context.go(AppRoute.splash.path);
+  }
 }
